Keep concatenated sequence numbers in signed arithmetic and cap confidence at 1
Negative query offsets (chunk-overlap) hit a negative double-to-uint cast in
ResultEntryConcatenator, which saturates to 0 on .NET Core and is unspecified
on .NET Framework, desyncing query sequence numbers from match times. The
shift is now computed in signed arithmetic, clamped so sequence numbers can
never wrap below zero. FindGaps compares sequence numbers in signed math for
the same reason. ConfidenceCalculator caps the result at 1 (paths extending
past the declared query/track length shrank the denominator below the
numerator) and guards against a non-positive denominator. Coverage now caches
gap computations, which consumers read repeatedly per candidate.

+++ src/SoundFingerprinting/Query/MatchedWithExtensions.cs
@@ -447,7 +447,8 @@ namespace SoundFingerprinting.Query
                 float startsAt = matches[i - 1].Item2;
                 float endsAt = matches[i].Item2;
                 float gap = (float)SubFingerprintsToSeconds.GapLengthToSeconds(endsAt, startsAt, fingerprintLength);
-                bool sequenceNumberIncremented = matches[i].Item1 - matches[i - 1].Item1 > 1;
+                // entries are ordered by time, not by sequence number, so the delta can be negative; unsigned subtraction would wrap
+                bool sequenceNumberIncremented = (long)matches[i].Item1 - matches[i - 1].Item1 > 1;
                 float start = endsAt - gap;
                 if (!(endsAt <= start) && gap > permittedGap && sequenceNumberIncremented)
                 {
